requirements.txt
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-mock>=3.10.0
langchain_mistralai
